requirements.txt
numpy
opencv-python
pyyaml
h5py
hdf5plugin
numba
seaborn
torch